refactor: move raid signup button handler
