+++ src/events/interactionCreate.ts
@@ -1,7 +1,8 @@
 import { Client, Events } from 'discord.js';
 import { SupabaseClient } from '@supabase/supabase-js';
 import { Command } from '../types';
-import { handleRaidCreateModal, handleRaidSignupButton } from '../commands/raid';
+import { handleRaidCreateModal } from '../commands/raid';
+import { handleRaidSignupButton } from '../utils/button-handlers';
 
 export default function registerInteractionCreate(client: Client, commands: Map<string, Command>, supabase: SupabaseClient) {
   client.on(Events.InteractionCreate, async (interaction) => {
